+++ src/libraries/System.Buffers/ref/Configurations.props
@@ -1,7 +1,6 @@
 ﻿<Project>
   <PropertyGroup>
     <BuildConfigurations>
-      netstandard2.0;
       $(NetCoreAppCurrent);
     </BuildConfigurations>
   </PropertyGroup>
