--- src/libraries/System.Reflection.Emit.ILGeneration/src/ApiCompatBaseline.uap10.0.16299.txt
@@ -1,4 +1,0 @@
-Compat issues with assembly System.Reflection.Emit.ILGeneration:
-CannotRemoveAttribute : Attribute 'System.Runtime.CompilerServices.IsReadOnlyAttribute' exists on 'System.Reflection.Emit.Label' in the contract but not the implementation.
-TypeCannotChangeClassification : Type 'System.Reflection.Emit.Label' is marked as readonly in the contract so it must also be marked readonly in the implementation.
-Total Issues: 2
