+++ src/libraries/System.Reflection.Emit.Lightweight/ref/Configurations.props
@@ -1,13 +1,7 @@
 ﻿<Project>
   <PropertyGroup>
-    <PackageConfigurations>
-      netstandard2.0;
-    </PackageConfigurations>
     <BuildConfigurations>
-      $(PackageConfigurations);
       $(NetCoreAppCurrent);
-      <!-- Use inbox facade on $(NetFrameworkCurrent) -->
-      _$(NetFrameworkCurrent);
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
