--- src/libraries/System.Reflection.TypeExtensions/pkg/System.Reflection.TypeExtensions.pkgproj
@@ -1,26 +1,0 @@
-﻿<Project DefaultTargets="Build">
-  <Import Project="$([MSBuild]::GetPathOfFileAbove(Directory.Build.props))" />
-  <ItemGroup>
-    <ProjectReference Include="..\ref\System.Reflection.TypeExtensions.csproj">
-      <SupportedFramework>net461</SupportedFramework>
-    </ProjectReference>
-    <ProjectReference Include="..\src\System.Reflection.TypeExtensions.csproj" />
-    <HarvestIncludePaths Include="ref/net46;lib/net46" />
-    <HarvestIncludePaths Include="lib/netcore50;runtimes/aot/lib/netcore50" />
-    <HarvestIncludePaths Include="ref/netstandard1.3" />
-    <HarvestIncludePaths Include="ref/netstandard1.5" />
-    <HarvestIncludePaths Include="lib/netcoreapp1.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <InboxOnTargetFramework Include="netcoreapp2.0" />
-    <InboxOnTargetFramework Include="uap10.0.16299" />
-    <File Include="$(PlaceholderFile)">
-      <TargetPath>runtimes/aot/lib/uap10.0.16299</TargetPath>
-    </File>
-    <InboxOnTargetFramework Include="$(AllXamarinFrameworks)" />
-    <!-- this package is part of the implementation closure of NETStandard.Library
-         therefore it cannot reference NETStandard.Library -->
-    <SuppressMetaPackage Include="NETStandard.Library" />
-  </ItemGroup>
-  <Import Project="$([MSBuild]::GetPathOfFileAbove(Directory.Build.targets))" />
-</Project>
