+++ src/libraries/System.Reflection.TypeExtensions/ref/Configurations.props
@@ -1,9 +1,7 @@
 ﻿<Project>
   <PropertyGroup>
     <BuildConfigurations>
-      netstandard2.0;
-      $(NetFrameworkCurrent);
-      net461;
+      $(NetCoreAppCurrent);
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
