+++ src/libraries/System.Reflection.TypeExtensions/src/Configurations.props
@@ -1,14 +1,6 @@
 ﻿<Project>
   <PropertyGroup>
-    <PackageConfigurations>
-      net461;
-      netstandard2.0;
-      netstandard1.3;
-      netstandard1.5;
-    </PackageConfigurations>
     <BuildConfigurations>
-      $(PackageConfigurations);
-      $(NetFrameworkCurrent);
       $(NetCoreAppCurrent)-Windows_NT;
       $(NetCoreAppCurrent)-Unix;
     </BuildConfigurations>
