+++ src/libraries/System.Reflection.TypeExtensions/tests/MetadataTokenTests.cs
@@ -11,6 +11,16 @@ namespace System.Reflection.Tests
 {
     public class MetadataTokenTests
     {
+        [Fact]
+        public void GetMetadataToken_ArraysAndTheirMembers_DoNotHaveMetadataTokens()
+        {
+            Assert.False(typeof(byte[]).GetTypeInfo().HasMetadataToken());
+            Assert.Throws<InvalidOperationException>(() => typeof(byte[]).GetTypeInfo().GetMetadataToken());
+
+            Assert.False(typeof(byte[]).GetMethods()[0].HasMetadataToken());
+            Assert.Throws<InvalidOperationException>(() => typeof(byte[]).GetTypeInfo().GetMetadataToken());
+        }
+
         private class Test<T> { }
 
         public int Field;
