+++ src/libraries/restore/netstandard/netstandard.depproj
@@ -53,6 +53,9 @@
     </PackageReference>
   </ItemGroup>
   <ItemGroup Condition="'$(IsNETStandard2x)' == 'true'">
+    <PackageReference Include="System.Buffers">
+      <Version>4.5.0</Version>
+    </PackageReference>
     <PackageReference Include="System.Security.Cryptography.Cng">
       <Version>4.5.0</Version>
     </PackageReference>
